backend/requirements-test.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (use sqlite for testing instead of psycopg2)
sqlalchemy==2.0.23
alembic==1.12.1

# Task Queue
celery[redis]==5.3.4
redis==5.0.1

# HTTP & API Clients
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Authentication & Authorization
authlib==1.3.0
python-jose[cryptography]==3.3.0
pyjwt==2.7.0
passlib[bcrypt]==1.7.4

# OAuth Providers
google-auth==2.25.2
requests-oauthlib==1.3.0

# Billing
stripe==7.4.0

# Web Scraping
beautifulsoup4==4.12.2

# Logging & Monitoring
python-dotenv==1.0.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
